Update opencl-amd.sh
Changed variable names to be a more similar format to the rest of the variables originally in place. Added improved update resiliency to regEx during repository fixing. 

I don't know what exactly the edited document would look like since I don't have a card newer than Polaris, but the updated `sed` pattern will always find and replace a line that contains the phrase `rhel`, a number (with the possibility of including a decimal or several), a forward slash, and another number and/or decimal with number. I tested this functionality substantially with various test strings to ensure it would work in the future. Included with my new code are comments explaining the what/why of what I've done.